+++ changelog.md
@@ -8,6 +8,7 @@
 ### Improvements
 - Player count and roster errors from TTS boxscore upload are passed to Discord instead of blocking the upload
 - Improved the #results message to link to thre game thread when possible
+- LFG Posts can now be cancelled by guild moderators
 
 ### Bug Fixes
 - Subscribing to a scheduling poll now notifies on every new "Yes" answer instead of just poll close.